requirements.txt
debugpy==1.8.15
requests==2.31.0
python-dateutil==2.9.0.post0
garminconnect==0.2.28
python-dotenv==1.1.1
flask==2.3.2
pytz==2025.2

# Google API Client Libraries
#google-api-python-client==2.108.0
#google-auth-httplib2==0.1.1
#google-auth-oauthlib==1.1.0
#google-auth==2.23.4


# HTTP and Authentication Support
#httplib2==0.22.0
#requests-oauthlib==1.3.1